Reestructuracion

+++ models.py
@@ -1,4 +1,3 @@
-# models.py
 from typing import Optional
 # models.py
 from typing import Optional, List
@@ -78,7 +77,7 @@ class Renta(SQLModel, table=True):
     id: Optional[int] = Field(default=None, primary_key=True)
     edificio: str
     habitaciones: int
-    banos: int
+    banos: float
     lat: float
     lon: float
     mascotas: bool
